# web/src/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for the Rack Gateway administration and proxy services.",
        "title": "Rack Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/settings/rack_tls_cert/refresh": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Fetches and stores the latest TLS certificate for the configured Convox rack.",
                "tags": [
                    "Settings"
                ],
                "summary": "Refresh rack TLS certificate",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/db.RackTLSCert"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "CSRFToken": {
                "description": "HMAC-derived CSRF token tied to the active session.",
                "type": "apiKey",
                "name": "X-CSRF-Token",
                "in": "header"
            },
            "SessionCookie": {
                "description": "HttpOnly session cookie issued after OAuth login.",
                "type": "apiKey",
                "name": "Cookie",
                "in": "header"
            }
        },
        "schemas": {
...
            "db.RackTLSCert": {
                "type": "object",
                "properties": {
                    "fetched_at": {
                        "type": "string"
                    },
                    "fingerprint": {
                        "type": "string"
                    },
                    "pem": {
                        "type": "string"
                    }
                }
            },
...
            "handlers.ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}